requirements.txt
flask
rasterio
shapely
pyproj
numpy
pillow
matplotlib